rename new_fn.rs -> new.rs
