+++ src/lib.rs
@@ -9,7 +9,7 @@ mod deserialize_time_range;
 mod fields;
 mod filter;
 mod load;
-mod new_fn;
+mod new;
 mod serde;
 mod store;
 
@@ -115,7 +115,7 @@ pub fn store(args: TokenStream, item: TokenStream) -> TokenStream {
     let load_and_delete = load::generate(&model, &timestamp, &group_by, &packed_name, &table_name);
     let decompress = decompress::generate(&model, &timestamp, &group_by, float_round, &table_name, using_chrono);
     let store_and_store_grouped = store::generate(&model, &timestamp, &group_by, float_round, &table_name);
-    let new_fn = new_fn::generate(&model, &timestamp, &group_by, float_round, using_chrono);
+    let new = new::generate(&model, &timestamp, &group_by, float_round, using_chrono);
     let serde = serde::generate();
 
     quote! {
@@ -131,7 +131,7 @@ pub fn store(args: TokenStream, item: TokenStream) -> TokenStream {
         }
 
         impl #packed_name {
-            #new_fn
+            #new
 
             #load_and_delete
